添加部分网络请求逻辑

+++ BoreyAdSDK/Splash/BoreySplashAdFiller.h
@@ -11,7 +11,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface BoreySplashAdFiller : NSObject
 
-- (void) fill: (NSString *) tagId bidFloor: (float) bidFloor;
+- (void) fill: (NSString *) tagId bidFloor: (long) bidFloor;
 
 @end
